+++ scrapy_playwright/headers.py
@@ -2,10 +2,11 @@
 This module includes functions to process request headers.
 Refer to the PLAYWRIGHT_PROCESS_REQUEST_HEADERS setting for more information.
 """
-
+import warnings
 from urllib.parse import urlparse
 
 from playwright.async_api import Request as PlaywrightRequest
+from scrapy.exceptions import ScrapyDeprecationWarning
 from scrapy.http.headers import Headers
 
 
@@ -17,21 +18,22 @@ async def use_scrapy_headers(
     """Scrapy headers take precedence over Playwright headers for navigation requests.
     For non-navigation requests, only User-Agent is taken from the Scrapy headers."""
 
-    headers = scrapy_headers.to_unicode_dict()
+    scrapy_headers_str = scrapy_headers.to_unicode_dict()
+    playwright_headers = await playwright_request.all_headers()
 
     # Scrapy's user agent has priority over Playwright's
-    headers.setdefault("user-agent", playwright_request.headers.get("user-agent"))
+    scrapy_headers_str.setdefault("user-agent", playwright_headers.get("user-agent"))
 
     if playwright_request.is_navigation_request():
         if browser_type == "firefox":
             # otherwise this fails with playwright.helper.Error: NS_ERROR_NET_RESET
-            headers["host"] = urlparse(playwright_request.url).netloc
-        return headers
+            scrapy_headers_str["host"] = urlparse(playwright_request.url).netloc
+        return scrapy_headers_str
 
     # override user agent, for consistency with other requests
-    if headers.get("user-agent"):
-        playwright_request.headers["user-agent"] = headers["user-agent"]
-    return playwright_request.headers
+    if scrapy_headers_str.get("user-agent"):
+        playwright_headers["user-agent"] = scrapy_headers_str["user-agent"]
+    return playwright_headers
 
 
 async def use_playwright_headers(
@@ -39,5 +41,11 @@ async def use_playwright_headers(
     playwright_request: PlaywrightRequest,
     scrapy_headers: Headers,
 ) -> dict:
-    """Return headers from the Playwright request, unaltered"""
-    return playwright_request.headers
+    warnings.warn(
+        "The 'scrapy_playwright.headers.use_playwright_headers' function is"
+        " deprecated, please set 'PLAYWRIGHT_PROCESS_REQUEST_HEADERS=None'"
+        " instead.",
+        category=ScrapyDeprecationWarning,
+        stacklevel=1,
+    )
+    return await playwright_request.all_headers()